Add knowledge maintainer service
This will search through the database for spatial things that are based on a foursquare resource that haven't been populated.  This will populate each of the unpopulated nodes with data from the database.

Update the utilities to populate the storage payload correctly
Update the foursquarelookup to correctly create the client (was missing the event parameter in the configuration loaded method).
Added the foursquarelookup and knowledge maintainer plugins to the bot.